refactor: move notion storage to its own package
