+++ internal/vacancy-worker/driven/storage/notion/model.go
@@ -1,4 +1,4 @@
-package storage
+package notion
 
 import (
 	"strings"
