+++ internal/vacancy-worker/driven/storage/notion/notion.go
@@ -1,4 +1,4 @@
-package storage
+package notion
 
 import (
 	"context"
