+++ internal/vacancy-worker/driven/storage/notion/notion_test.go
@@ -1,4 +1,4 @@
-package storage_test
+package notion_test
 
 import (
 	"context"
@@ -6,7 +6,7 @@ import (
 
 	"github.com/ghazlabs/idn-remote-entry/internal/shared/core"
 	"github.com/ghazlabs/idn-remote-entry/internal/testutil"
-	"github.com/ghazlabs/idn-remote-entry/internal/vacancy-worker/driven/storage"
+	storage "github.com/ghazlabs/idn-remote-entry/internal/vacancy-worker/driven/storage/notion"
 	"github.com/go-resty/resty/v2"
 	"github.com/riandyrn/go-env"
 	"github.com/stretchr/testify/require"
